+++ src/utils/UpdateChecker.js
